requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
arabic-reshaper==3.0.0
asn1crypto==1.5.1
attrs==26.1.0
azure-common==1.1.28
azure-core==1.39.0
azure-identity==1.25.3
azure-search-documents==11.6.0
beautifulsoup4==4.14.3
blis==1.3.3
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cloudpathlib==0.23.0
colorama==0.4.6
confection==1.3.3
cryptography==46.0.7
cssselect2==0.9.0
cymem==2.0.13
dataclasses-json==0.6.7
distro==1.9.0
emoji==2.15.0
en_core_web_sm==3.8.0
et_xmlfile==2.0.0
filelock==3.29.0
filetype==1.2.0
freetype-py==2.5.1
frozenlist==1.8.0
greenlet==3.4.0
grpcio==1.80.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
installer==0.7.0
isodate==0.7.2
Jinja2==3.1.6
jiter==0.14.0
jsonpatch==1.33
jsonpointer==3.1.1
langchain==1.2.15
langchain-classic==1.0.4
langchain-community==0.4.1
langchain-core==1.3.1
langchain-openai==1.1.14
langchain-qdrant==1.1.0
langchain-text-splitters==1.1.2
langdetect==1.0.9
langgraph==1.1.8
langgraph-checkpoint==4.0.2
langgraph-prebuilt==1.0.10
langgraph-sdk==0.3.13
langsmith==0.7.33
llvmlite==0.47.0
lxml==6.1.0
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
msal==1.36.0
msal-extensions==1.3.1
multidict==6.7.1
murmurhash==1.0.15
mypy_extensions==1.1.0
networkx==3.6.1
numba==0.65.0
numpy==2.4.4
olefile==0.47
openai==2.32.0
openpyxl==3.1.5
orjson==3.11.8
ormsgpack==1.12.2
oscrypto==1.3.0
packaging==26.1
pandas==3.0.2
pillow==12.2.0
portalocker==3.2.0
preshed==3.0.13
propcache==0.4.1
protobuf==7.34.1
psutil==7.2.2
pycairo==1.29.0
pycparser==3.0
pydantic==2.13.3
pydantic-settings==2.14.0
pydantic_core==2.46.3
Pygments==2.20.0
pyHanko==0.34.1
pyhanko-certvalidator==0.30.2
PyJWT==2.12.1
PyMuPDF==1.27.2.2
pypdf==6.10.2
pypdfium2==5.7.1
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
python-iso639==2026.4.20
python-magic==0.4.27
python-oxmsg==0.0.2
pywin32==311
PyYAML==6.0.3
qdrant-client==1.17.1
RapidFuzz==3.14.5
regex==2026.4.4
reportlab==4.4.10
requests==2.33.1
requests-toolbelt==1.0.0
rich==15.0.0
rlPyCairo==0.4.0
shellingham==1.5.4
six==1.17.0
smart_open==7.6.0
sniffio==1.3.1
soupsieve==2.8.3
spacy==3.8.14
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.49
srsly==2.5.3
svglib==1.6.0
tenacity==9.1.4
thinc==8.3.13
tiktoken==0.12.0
tinycss2==1.5.1
tqdm==4.67.3
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
unstructured==0.22.22
unstructured-client==0.43.2
uritools==6.0.2
urllib3==2.6.3
uuid_utils==0.14.1
wasabi==1.1.3
watchdog==6.0.0
weasel==1.0.0
webencodings==0.5.1
wrapt==2.1.2
xhtml2pdf==0.2.17
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
